add eslint add prettier integrate concept of runner libs to share tasks
