+++ tests/k6/.eslintrc
@@ -1,4 +1,8 @@
 {
+  "plugins": [
+    "simple-import-sort",
+    "import"
+  ],
   "parser": "@typescript-eslint/parser",
   "parserOptions": {
     "ecmaVersion": 2020,
@@ -10,5 +14,10 @@
     "plugin:prettier/recommended"
   ],
   "rules": {
+    "simple-import-sort/imports": "error",
+    "simple-import-sort/exports": "error",
+    "import/first": "error",
+    "import/newline-after-import": "error",
+    "import/no-duplicates": "error",
   }
 }
