+++ tests/k6/package.json
@@ -36,8 +36,10 @@
     "babel-plugin-lodash": "^3.3.4",
     "eslint": "^7.14.0",
     "eslint-config-prettier": "^6.15.0",
+    "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jest": "^23.8.2",
     "eslint-plugin-prettier": "^3.2.0",
+    "eslint-plugin-simple-import-sort": "^6.0.1",
     "husky": "^4.3.0",
     "jest": "^25.4.0",
     "k6": "^0.0.0",
