+++ tests/k6/src/lib/index.ts
@@ -1,6 +1,6 @@
 export * as api from './api';
-export * as playbook from './playbook';
-export * as auth from './auth';
+export { default as auth } from './auth';
 export * as defaults from './defaults';
+export * as playbook from './playbook';
 export * as types from './types';
 export * as utils from './utils';
